Add files via upload
Hi!
This work contains an algorithm, code, and output of the program that displays on a 2D object the minimum value from the user's input. The minimum point will be located inside this object (if inputs are not negative, otherwise, the point will be beyond it).
Location of min value point will depend on the maximum input value, visually representing the difference between the maximum and minimum inputs, and changing from input to input.