Add files via upload
This programme is a simple to-do list manager where the user can create a new file to store tasks or select one of the existing to-do list files to work with. In addition, the user can add new tasks to the selected list. The user can also view all the tasks in the selected list. Likewise, the user can modify an existing task. And finally the user can remove a task from the list when, for example, he/she completes it. 
The programme offers a user-friendly command line interface for task management, providing the ability to work with multiple to-do lists.